requirements.txt
fastapi
uvicorn
python-socketio
python-multipart
# Google GenAI SDK (v1beta)
google-genai
# Computer Vision & Audio
opencv-python
pyaudio
pillow
mss
# Browser Automation
playwright
# Utilities
python-dotenv
# Face & Hand tracking
mediapipe
# Long-term Memory
supermemory
